pyproject.toml
[project]
name = "comfyui-chutes-nodes"
version = "1.0.0"
description = "ComfyUI custom nodes for Chutes AI image generation API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
]

[project.urls]
Homepage = "https://github.com/minipasila/chutes-comfyui-node"
Issues = "https://github.com/minipaila/chutes-comfyui-node/issues"

